create services to handle with pictures, now have service to create, delete and get a picture - was create to the service to update image product and get product data specify

+++ src/Application/Products/RemoveProduct/RemoveProductService.cs
@@ -1,18 +1,30 @@
 ﻿using Application.Common.ServiceHandler;
+using Application.Picturies.RemovePicture;
 using Contracts.Repositories;
 using Contracts.Works;
 using Domain.Common.Errors;
+using Domain.Products;
 using ErrorOr;
 
 namespace Application.Products.RemoveProduct;
 
-public class RemoveProductService(IProductRepository productRepository, IUnitOfWork unitOfWork) : IServiceHandler<RemoveProductRequest, ErrorOr<Success>>
+public class RemoveProductService(
+    IProductRepository productRepository,
+    IRemovePictureService removePictureService,
+    IUnitOfWork unitOfWork) : IServiceHandler<RemoveProductRequest, ErrorOr<Success>>
 {
     public async Task<ErrorOr<Success>> Handler(RemoveProductRequest request)
     {
         var product = await productRepository.GetById(request.ProductId);
 
         if (product is null) return Errors.Product.ProductNotFound;
+
+        var result = await removePictureService.Handler(
+            new RemovePictureRequest(
+                nameof(Product),
+                product.Id));
+
+        if (result.IsError) return result.Errors;
 
         productRepository.Remove(product);
