+++ src/Application/Users/RemoveAccount/RemoveAccountService.cs
@@ -4,6 +4,9 @@ using Domain.Common.Errors;
 using Contracts.Repositories;
 using Contracts.Works;
 using ErrorOr;
+using Application.Picturies.RemovePicture;
+using Domain.Companies;
+using Domain.Users;
 
 namespace Application.Users.RemoveAccount;
 
@@ -11,12 +14,14 @@ public class RemoveAccountService(
     IPasswordHasher passwordHasher,
     IUserAuthenticated userAuthenticated,
     IUnitOfWork unitOfWork,
-    IUserRepository userRepository)
+    IUserRepository userRepository,
+    IRemovePictureService removePictureService)
     : IServiceHandler<RemoveAccountRequest, ErrorOr<Success>>
 {
     public async Task<ErrorOr<Success>> Handler(RemoveAccountRequest request)
     {
         var userId = userAuthenticated.GetUserId();
+        var companyId = userAuthenticated.GetCompanyId();
 
         var user = await userRepository.GetById(userId);
 
@@ -29,6 +34,20 @@ public class RemoveAccountService(
         {
             return Errors.Authentication.InvalidCredentials;
         }
+
+        var resultRemovePictureUser = await removePictureService.Handler(
+            new RemovePictureRequest(
+                nameof(User),
+                user.Id));
+
+        if (resultRemovePictureUser.IsError) return resultRemovePictureUser.Errors;
+
+        var resultRemovePictureCompany = await removePictureService.Handler(
+            new RemovePictureRequest(
+                nameof(Company),
+                companyId));
+
+        if (resultRemovePictureCompany.IsError) return resultRemovePictureCompany.Errors;
 
         userRepository.Remove(user);
